pyproject.toml
[project]
name = "vision-mouse"
version = "1.0.0"
description = "Offline hand gesture mouse controller for Windows"
requires-python = ">=3.11"

dependencies = [
    "mediapipe>=0.10.32",
    "opencv-python>=4.10.0",
    "pynput>=1.7.6",
    "keyboard>=0.13.5",
    "customtkinter>=5.2.2",
    "pystray>=0.19.5",
    "pillow>=10.4.0",
    "pyinstaller>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vision_mouse"]
